pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cryptex"
version = "0.1.0"
description = "Minimal crypto trading pipeline for Binance"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [{name = "CryptEx"}]
keywords = ["crypto", "trading", "binance", "trading-bot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "asyncclick>=8.0.0",
    "httpx>=0.27.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.scripts]
cryptex = "cryptex.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"

[tool.setuptools.packages.find]
where = ["."]
include = ["cryptex*"]
